add flag which allows specification of output format (#68)
This PR adds `--shell` to `keyconjurer get`. `--shell` allows a user to determine
which output format will be used to output credentials to standard output if `--out`
is set to `env`.

This is useful for users on Windows running KeyConjurer in WSL;
KeyConjurer determines the OS as Windows and so outputs variables in
Windows' BASIC format, which cannot be executed in a subshell by
default.

The options for this new flag are as follows:

* `basic` - Output credentials in Windows' command prompt format, BASIC.
* `bash` - Output credentials in a format that all Bash and Bash-alike
shells may use.
* `powershell` - Output credentials in PowerShell format.
* `infer` - KeyConjurer determines which shell you are using and then
outputs credentials in a format for that shell. This was the previous
behavior. If the shell cannot be determined, then the output will be
formatted as if the user had selected `bash`.

If `--shell` is omitted, then `infer` will be chosen by default, which
matches the previous behaviour. Example usage:

  keyconjurer get my-account --shell bash