Fix a bug where account selection was not updated internally
If the list of accounts changed, the browser would default to using the
first element in the list but it would not update the internal state of
the Form.

This occurred because browsers will visually default to, and select, the
first <option> in a <select> - indeed, interrogating the `value`
property of the <select> as done in tests shows that this works as
expected!

However, we provide the internal state of the KeyRequestForm component
as the value to the form, and that will be undefined until the user
interacts with the dropdown (until now).

Resolves #18