Add Duo OIDC redirect flow support for second factor (#91)
This change adds support to KeyConjurer for users who have configured Duo as an OIDC IdP in Okta.

It uses non-public, undocumented APIs of Okta to achieve this. This is not a recommended approach and is introduced as a stop-gap. You should consider configuring Duo as a second factor within Okta instead of using it as an OIDC factor. However, as best as the author can tell, this is the only way of getting the Duo OIDC IdP working programmatically.

Please don't be discouraged by the large number of changes - most of the changes are in the vendor/ folder, which is excluded from the linguist preview in the Files Changed view, but is not excluded from the diff stats by Github.
How does it work?

A high level summary: The Okta flow has been reworked to use applications to kick off the authorization flow, rather than using the /api/v1/authn endpoint. KeyConjurer will visit the application link for an application with a certain ID, and will interact with the web page just like the browser might. The old flow looked something like this:

1. User submits credentials and application ID.
2. KeyConjurer verifies that the users credentials are correct and initiates a session using /api/v1/authn.
3. After initiating a session, KeyConjurer initiates a two-factor prompt with Duo, if the user has a factor type of web and vendor of Duo, and goes through the Duo flow as described by Duo's website, using HTML parsing to extract values and post them back as appropriate.
4. After the Duo session is complete, the challenge response from Duo is sent back to Okta, which Okta uses to upgrade the users session to have MFA enabled.
5. KeyConjurer then visits the application link, manually adding cookies from the Okta session into the request. The application link will respond with a SAML response, which is extracted and handed to sts.AssumeRoleWithSAML.

The new flow looks something like this:

1. User submits credentials and application ID.
2. KeyConjurer retrieves the application link for the given application, returning an error if the application was not found.
3. A cookiejar is created to capture all cookies for a given session.
4. KeyConjurer visits this application link, and using regex, extracts the Okta stateToken from the embedded javascript within the page.
5. The username and password is POSTED to the /idp/idx/identify endpoint, which upgrades the session to know "who" the user is and indicates what second factor to use.
  5a. If the response indicates the user has an OIDC 2fa, the DuoFrameless flow is used.
  5b. If the response indicates the user has a challenge-authenticator 2fa type, the DuoIframe (legacy flow) is used.
  5c. Any other response will result in an error indicating there is no supported 2nd factor.
6. After the appropriate 2fa path is handled, the SAML assertion is received by sending a couple of more requests to Okta.

The implementation detail of both DuoFrameless and DuoIframe are very similar, with the primary difference being that the challenge response from DuoIframe is located within an iframe, whereas in DuoFrameless the challenge response is never exposed to KeyConjurer and is instead handled by redirects.

There is sufficent overlap between DuoFrameless and DuoIframe (and their respective implementations in the duo package) that the code could be deduplicated, but that is left for a future task and is not a priority right now.