Make it harder to cause a race condition to begin with
Instead of returning a struct with mutable struct, we can make it harder
to cause a race condition by wrapping all of that state instead of a
function closure which cannot be modified from the outside, and by using
sync.Once() to ensure that only one request is ever handled.

We only need to have the sync.Once() for close()ing the channel to
permit a user to permit a user to free the resources manually. It is
still possible to cause a panic by closing this channel before receiving
a request.