Remove Keyconjurer Upgrade
This feature has caused us a lot of problems over the years:

* It would brick itself on Windows if you failed to download the file
  for any reason
* Sometimes the DownloadURL would not be set
* Relying on the releases always being at a specific URL meant that it
  was difficult to change it for any reason (For example, to
support a stable and development release channel)

We are removing this command and will instead move to supporting package
managers like Homebrew for OSX and Chocolatey for Windows.

It is recommended that folks who depend on this project do the same.